Update config.mak
Update Makefile

move prefix and libdir into config.mak

Update Makefile

config.mak and litecross

removes arch specific from of litecross. We provide arch specific flags taken from Alpine Linux. This previous method was to apply changes to generic triples used when building like riscv64-linux-musl but when we do this a riscv64specific set of flags are passed as well. So these if are redundant for this modified version of mcm.

Updated config.mak.